Refactoring to reduce memory usage - main culprit is use of String. Using sprintf instead for concat etc

+++ pedalboard/src/Domain/button_event.h
@@ -11,15 +11,15 @@ struct ButtonEvent : public Some {
         None
     };
 
-    static const unsigned int LONG_HOLD_LIMIT_MS = 1500;
-    static const unsigned int LONG_HOLD_START_MS = 500;
+    static const unsigned short LONG_HOLD_LIMIT_MS = 1500;
+    static const unsigned short LONG_HOLD_START_MS = 500;
 
     button_event_type EventType;
 
     /*
      * The time a button was held for, only relevant for Hold events
      */
-    unsigned long HeldMs;
+    unsigned short HeldMs;
 
     /*
      * The event time for Press,LongPress is when the button is released.
